makefile a gerer ft_printf undefined

+++ libft/src/ft_printf/ft_putnbr_base.c
@@ -6,7 +6,7 @@
 /*   By: norabino <norabino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 14:44:09 by norabino          #+#    #+#             */
-/*   Updated: 2025/01/09 15:48:00 by norabino         ###   ########.fr       */
+/*   Updated: 2025/01/14 09:29:11 by norabino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	ft_putnbr_base(long nb, char *base, int *ptr_len)
 		nb = -nb;
 		*ptr_len += ft_print_c('-');
 	}
-	if (nb > ft_strlen(base) - 1)
+	if (nb > (long int)ft_strlen(base) - 1)
 		ft_putnbr_base(nb / ft_strlen(base), base, ptr_len);
 	*ptr_len += ft_print_c(base[nb % (unsigned int)ft_strlen(base)]);
 }
